# switch underlying interpoliation to use rates instead of discount factor